google-cloud/b24-spa-1106-sync/b24-spa-1106-sync-worker/requirements.txt
functions-framework==3.*
cloudevents>=1.9.0
requests>=2.31.0
google-cloud-secret-manager>=2.16.0
